fdsg

+++ Control/NumericBox/NumericBox.h
@@ -4,7 +4,7 @@
 
 struct PlusListener : public MouseListener{
 	PlusListener(Control& c) : _c(c){}
-	void mousePressed(Button& b, int x, int y, bool isLeft){
+	void mousePressed(Button& b, short x, short y, bool isLeft) {
 		NumericBox& tmp = static_cast<NumericBox&>(_c);
 		tmp.setValue(tmp.getValue() + 1);
 	}
@@ -14,7 +14,7 @@ private:
 
 struct MinusListener : public MouseListener{
 	MinusListener(Control& c) : _c(c){}
-	void mousePressed(Button& b, int x, int y, bool isLeft){
+	void mousePressed(Button& b, short x, short y, bool isLeft){
 		NumericBox& tmp = static_cast<NumericBox&>(_c);
 		tmp.setValue(tmp.getValue() - 1);
 	}
@@ -33,5 +33,6 @@ public:
 	void setValue(int val);
 	int getValue();
 	~NumericBox();
+	
 };
